# api/message/api/swagger.yaml
openapi: 3.0.0
info:
  title: Мессенджер
  description: |
    API Мессенджера - набор методов для получения списка чатов пользователя на Авито, получения сообщений в чате, отправки сообщения в чат и другие Через API Мессенджера можно организовать интеграцию между мессенджером Авито и сторонней системой в обе стороны

    **Авито API для бизнеса предоставляется согласно [Условиям использования](https://api.avito.ru/docs/public/APITermsOfServiceV1.pdf).**
  contact:
    email: supportautoload@avito.ru
  version: "1"
servers:
- url: https://api.avito.ru/
...
paths:
...
  /messenger/v1/webhook:
    post:
      tags:
      - Messenger
      summary: Включение уведомлений (webhooks)
      description: |
        Включение уведомлений
      operationId: postWebhook
      parameters:
      - name: Authorization
        in: header
        description: Токен для авторизации
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: Bearer ACCESS_TOKEN
      requestBody:
        description: Url на который будут отправляться уведомления
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/webhookSubscribeRequestBody'
      responses:
        "200":
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
          x-examples:
            application/json:
              ok: true
      security:
      - AuthorizationCode:
        - messenger:read
      - ClientCredentials: []
      x-code-samples:
      - lang: PHP
        source: "$ch = curl_init(\"https://api.avito.ru/messenger/v1/webhook\");\n\
          curl_setopt($ch, CURLOPT_HTTPHEADER, [\n\t\"Content-Type: application/json\"\
          ,\n\t\"Authorization: Bearer <TOKEN>\",\n]);\ncurl_setopt($ch, CURLOPT_POST,\
          \ true);\n\ncurl_setopt($ch, CURLOPT_POSTFIELDS, json_encode([\n\t\"url\"\
          \ => \"https://test.test\",\n], JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES));\n\
          \ncurl_setopt($ch, CURLOPT_RETURNTRANSFER, true);\n$result = curl_exec($ch);\n\
          curl_close($ch);"
...
components:
  schemas:
...
    webhookSubscribeRequestBody:
      required:
      - url
      type: object
      properties:
        url:
          type: string
          description: Url на который будут отправляться нотификации
...
    inline_response_200_1:
      type: object
      properties:
        ok:
          type: boolean
          example: true
      example:
        ok: true
...
  securitySchemes:
    AuthorizationCode:
      type: oauth2
      description: "Это API использует OAuth 2 с механизмом authorization_code. Ис\
        пользуйте его для доступа к данным других пользователей при разработке сто\
        роннего приложения. [Подробнее](/api-catalog/auth/documentation#tag/ApplicationAccess)"
      flows:
        authorizationCode:
          authorizationUrl: https://avito.ru/oauth
          tokenUrl: https://api.avito.ru/token
          scopes:
            autoload:reports: Получение отчетов Автозагрузки
            items:apply_vas: Применение дополнительных услуг
            items:info: Получение информации об объявлениях
            job:cv: Получение информации резюме
            job:negotiations: Получение информации об откликах на вакансии
            job:write: Изменение объявлений вертикали Работа
            messenger:read: Чтение сообщений в мессенджере Авито
            messenger:write: Модифицирование сообщений в мессенджере Авито
            short_term_rent:read: Получение информации об объявлениях краткосрочной
              аренды
            short_term_rent:write: Изменение объявлений краткосрочной аренды
            stats:read: Получение статистики объявлений
            user:read: Получение информации о пользователе
            user_balance:read: Получение баланса пользователя
            user_operations:read: Получение истории операций пользователя
    ClientCredentials:
      type: oauth2
      description: "Это API использует OAuth 2 с механизмом client_credentials. Ис\
        пользуйте его для доступа к возможностям своей личной учетной записи. [Под\
        робнее](#tag/Access)"
      flows:
        clientCredentials:
          tokenUrl: https://api.avito.ru/token
          scopes: {}